multiprocessing: use CriticalLock in a few places
This uses _thread.CriticalLock to prevent garbage collection while
modifying the finalizer registry and in the resource tracker.

There is a potential deadlock when checking if the resource tracker is
running. This exists in upstream CPython, but occurs much more frequently
with biased reference counting. The call to _check_alive can trigger a
garbage collection that collects a multiprocessing lock. The finalizer
for that lock calls back into resource tracker leading to a deadlock.

This is rare in upstream CPython because most function calls do not
trigger a garbage collection.